Adding pagination for customer lists
Right now our app will only take up to the first 100 users in any
customer list. The customer list endpoint is now paginated, which is
what’s causing the problem.  This attempts to fix that problem by
essentially iterating through all pages.

It may be worthwhile for us to build in an artificial limit - in other
words, only take the first 2000 in any list.